Move Searchbar functionality into custom hook (#15)
